wtGallery/requirements.txt
asgiref==3.2.10
boto3==1.14.26
botocore==1.17.26
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-imagekit==4.0.2
django-leaflet==0.27.0
django-starcross-gallery==1.0.8
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
firebase-admin==4.3.0
gcloud==0.17.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
jws==0.1.3
msgpack==1.0.0
oauth2client==3.0.0
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
protobuf==4.0.0rc2
psycopg2==2.7.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
